Updat class simple use

+++ Math/libMath/Geometry/Implem/Diamond.cs
@@ -4,24 +4,24 @@ namespace Utopia.Math.Geometry.Implem
 {
     class Diamond : IArea
     {
-        public float D1 { get; set; }
-        public float D2 { get; set; }
+        public double D1 { get; set; }
+        public double D2 { get; set; }
 
         public Diamond()
         {
-            D1 = float.NaN;
-            D2 = float.NaN;
+            D1 = double.NaN;
+            D2 = double.NaN;
         }
 
-        public Diamond(float d1, float d2)
+        public Diamond(double d1, double d2)
         {
             D1 = d1;
             D2 = d2;
         }
 
-        public float ComputeArea()
+        public double ComputeArea()
         {
-            return ((float)((float)(D1 * D2) / 2.0));
+            return (D1 * D2) / 2.0;
         }
     }
 }
