+++ Math/libMath/Geometry/Implem/Rectangle.cs
@@ -4,22 +4,22 @@ namespace Utopia.Math.Geometry.Implem
 {
     class Rectangle : IArea
     {
-        public float A { get; set; }
-        public float B { get; set; }
+        public double A { get; set; }
+        public double B { get; set; }
 
         public Rectangle()
         {
-            A = 0.0f;
-            B = 0.0f;
+            A = 0.0;
+            B = 0.0;
         }
 
-        public Rectangle(float a, float b)
+        public Rectangle(double a, double b)
         {
             A = a;
             B = b;
         }
 
-        public float ComputeArea()
+        public double ComputeArea()
         {
             return A * B;
         }
